Move TReleaseDevicesActor into separate file
