+++ cloud/blockstore/libs/storage/volume/actors/ya.make
@@ -3,6 +3,7 @@ LIBRARY()
 SRCS(
     forward_write_and_mark_used.cpp
     read_disk_registry_based_overlay.cpp
+    release_devices_actor.cpp
     shadow_disk_actor.cpp
 )
